feat: 총 댓글, 좋아요 갯수를 조회하도록 수정하라

+++ src/main/java/prgrms/project/stuti/domain/feed/service/dto/PostDto.java
@@ -14,8 +14,8 @@ public record PostDto(
 	String contents,
 	String postImageUrl,
 	LocalDateTime createdAt,
-	int totalLikes,
-	int totalComments,
+	Long totalLikes,
+	Long totalComments,
 	boolean isliked
 ) {
 	@Builder
